+++ api/index.d.ts
@@ -1,0 +1,9 @@
+declare global {
+  namespace Express {
+    interface Locals {
+      username: string;
+    }
+  }
+}
+
+export {};
