+++ api/src/controllers/auth.ts
@@ -2,6 +2,7 @@ import type { Request, Response } from "express";
 import User from "../models/user.js";
 import { createAuthToken, decodeAuthToken } from "../utils/authToken.js";
 import { z } from "zod";
+import { splitInHalf } from "../utils/index.js";
 
 const usernameAndPasswordSchema = z.object({
   username: z.string().min(6).max(30).trim(),
@@ -50,36 +51,27 @@ export async function login(req: Request, res: Response) {
 
   const authToken = createAuthToken(username);
 
-  res.cookie("authToken", authToken, {
+  const [firstHalf, secondHalf] = splitInHalf(authToken);
+
+  res.cookie("authToken1", firstHalf, {
     maxAge: 7 * 24 * 60 * 60 * 1000, // 7 days
     httpOnly: true,
     // secure: true, // TODO: set to true if in production env
     sameSite: "lax",
   });
 
+  res.cookie("authToken2", secondHalf, {
+    maxAge: 7 * 24 * 60 * 60 * 1000, // 7 days
+    httpOnly: false,
+    // secure: true, // TODO: set to true if in production env
+    sameSite: "lax",
+  });
+
   return res.status(200).json({ authenticated: true });
 }
-//checks for the authorization token in the cookies, the endpoint /auth needs to specifically be called. 
+
 export async function isAuthorized(req: Request, res: Response) {
-  const validation = authTokenSchema.safeParse(req.cookies.authToken);
-  if (!validation.success) {
-    return res.status(401).json({ error: "Not Authorized" });
-  }
-  const authToken = validation.data;
-
-  const payload = decodeAuthToken(authToken);
-
-  if (!payload) {
-    return res.status(401).json({ error: "Not Authorized" });
-  }
-
-  const { username } = payload;
-
-  const user = await User.findByUsername(username);
-
-  if (!user) {
-    return res.status(401).json({ error: "Not Authorized" });
-  }
+  const { username } = res.locals;
 
   return res.status(200).json({ authorized: true, username });
 }
